Added support for headless env
The draw.io application requires a graphical display, which does not exist in headless environments, such as a Docker container.

This commit implements a workaround.

Added the 'xvfb_wrapper.R' file, which contains 2 methods for headless environments.
* 'is.headless.env()' detects whether the current env is headless, i.e., without graphical display.
  * We use either 'xrandr' or the env variable 'DISPLAY' to do so. It should work correctly on Linux.
  * However, Windows and Mac OS X will probably not detect correctly whether an environment is headless.
* 'wrap.xvfb' wraps the call to draw.io in a call to 'xvfb-run', which will provide a virtual display so that draw.io can work correctly.

The 'drawio.engine' was slightly modified to take these new methods into account.